backend/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.2
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyJWT==2.10.1
email-validator==2.3.0